feat: opt docs for Actions
