+++ components/actions/__tests__/demo-extend.test.ts
@@ -1,0 +1,3 @@
+import { extendTest } from '../../../tests/shared/demoTest';
+
+extendTest('prompts');
