+++ components/actions/__tests__/demo.test.ts
@@ -1,0 +1,3 @@
+import demoTest from '../../../tests/shared/demoTest';
+
+demoTest('prompts');
