+++ components/actions/__tests__/image.test.ts
@@ -1,0 +1,5 @@
+import { imageDemoTest } from '../../../tests/shared/imageTest';
+
+describe('prompts image', () => {
+  imageDemoTest('prompts');
+});
