+++ components/actions/demo/sub.md
@@ -1,7 +1,7 @@
 ## zh-CN
 
-更多菜单项。
+通过设置 `children` 属性来展示更多菜单项，含有 `children` 的项可以不配置 icon，会默认使用 `<EllipsisOutlined />`
 
 ## en-US
 
-More menu items.
+Display more menu items by setting the `children` property. Items with `children` don't need to configure an icon; `<EllipsisOutlined />` will be used by default.
